Massvie overhaul
This must be like v3.0 of my build out. It's now using Webpack for building, has the capabuility to run a routable express server, one that can be used for api end points and  asset serving. The piece that took a long time was getting React to work well with the server. But I've figured out an organizational strucutre that allows a developer to build and deploy multiple react apps, each one with it's own webpack endpoint and shared bundles.  I've always had a problem with the idea that the React ecosystem seemed to expect you'd only ever have a single SPA to rule them all, so now I don't have to.