fixed drawer edge case
Edge case: when someone uses the drawer button code sets an linline style attribute on the element. so if user resizes, the media queries no longer wored because the inline style always won, so did a clean up on window resize to remove any inline style that might be present to allow media queries to work.